Move TwitterMediaExtensionFilter around
